server/requirements.txt
chromadb
google-generativeai
sentence-transformers
pyTelegramBotAPI
flask
